Module swaystatus::blocks::prelude

Re-exports

pub use super::BlockEvent;

pub use super::CommonApi;

pub use crate::click::MouseButton;

pub use crate::errors::Error;

pub use crate::errors::OptionExt;

pub use crate::errors::Result;

pub use crate::errors::ResultExt;

pub use crate::formatting::config::Config as FormatConfig;

pub use crate::formatting::value::Value;

pub use crate::widget::Spacing;

pub use crate::widget::State;

pub use crate::widget::Widget;

pub use crate::REQWEST_CLIENT;

Structs

Duration

A Duration type to represent a span of time, typically used for system timeouts.

StdString

A UTF-8–encoded, growable string.

Enums

StdResult

Result is a type that represents either success (Ok) or failure (Err).

Traits

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

StdError

Error is a trait representing the basic expectations for error values, i.e., values of type E in Result<T, E>.

Write

A trait for writing or formatting into Unicode-accepting buffers or streams.

Type Definitions

String

A convenience alias for a [LazyCompact] layout [SmartString].

Derive Macros

Deserialize